split qwen36 moe legacy decode helper
